backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==45.0.2
defusedxml==0.7.1
Django==5.1.7
django-bootstrap5==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-extra-fields==3.7.0
filelock==3.16.1
filetype==1.2.0
fsspec==2024.12.0
idna==3.10
iniconfig==2.0.0
Jinja2==3.1.5
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
oauthlib==3.2.2
packaging==24.2
pandas==2.2.3
pillow==11.2.1
pluggy==1.5.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.16.1
PySide6==6.8.1
PySide6_Addons==6.8.1
PySide6_Essentials==6.8.1
pyTelegramBotAPI==4.26.0
pytest==8.3.5
pytest-django==4.5.2
pytest-lazy-fixture==0.6.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
pytils==0.4.1
pytz==2025.1
requests==2.32.3
requests-oauthlib==2.0.0
setuptools==75.8.0
shiboken6==6.8.1
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.6.1
sqlparse==0.5.3
sympy==1.13.1
telebot==0.0.5
torch==2.5.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.4.0
